Extract window queries tests to separate class
Cherry-pick of https://github.com/prestosql/presto/commit/fcc18ff1c5360efdc43ac40629ad2845065de97c
